+++ src/orders/index.ts
@@ -165,9 +165,9 @@ export interface IOrdersService {
   ): Promise<AcceptQuoteResponse>;
 
   /**
-   * Edit Order (Beta)
+   * Edit Order
    *
-   * Edit an open order. This feature is in beta please reach out to your Coinbase Prime account manager for more information.
+   * Edit an open order.
    */
   editOrder(
     request: EditOrderRequest,