+++ src/services.ts
@@ -50,6 +50,8 @@ export { WalletsService, IWalletsService } from './wallets';
 // Export model types and enums
 export type * from './model/';
 export * from './model/enums/';
+export type * from './model/errors';
+export * from './model/errors/enums';
 
 // Export only the types that services actually need
 export type {